Bump to latest firebase messaging
