+++ gplay.gradle
@@ -1,3 +1,3 @@
 dependencies {
-    implementation "com.google.firebase:firebase-messaging:17.3.0"
+    implementation "com.google.firebase:firebase-messaging:20.1.3"
 }
